Move some utils to `lib` and fix circular deps
