+++ src/components/colors/ColorBlock.tsx
@@ -7,7 +7,7 @@ import {
 import { CheckIcon, CopyIcon } from 'lucide-react';
 
 import { Tooltip } from '@/components/ui/Tooltip';
-import { isHexColorLight, isValidHexColor } from '@/utils';
+import { isHexColorLight, isValidHexColor } from '@/lib/colorUtils';
 import { cn } from '@/lib/utils';
 
 export type ColorBlockProps = ComponentPropsWithoutRef<'button'> & {
