+++ src/lib/calculateSteps.ts
