--- src/types/index.ts
@@ -1,4 +1,0 @@
-export type * from './codeGen';
-export type * from './colorSuggestions';
-export type * from './globalState';
-export type * from './utils';
