--- src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './calculateSteps';
-export * from './codeGen';
-export * from './colorUtils';
